fix -fpermissive warning

+++ Inc/HALALMock/Models/PinModel/Pin.hpp
@@ -165,7 +165,7 @@ struct hash<Pin> {
         using std::string;
 
         return ((hash<uint16_t>()(k.gpio_pin) ^
-                 (hash<uint32_t>()((uint32_t)(k.port)) << 1)) >>
+                 (hash<uint32_t>()((uint32_t)((uintptr_t)k.port)) << 1)) >>
                 1);
     }
 };
